requirements.txt
flask
requests
beautifulsoup4
googlesearch-python